feat: build header for MVP. header is sticky. header is responsive.
fixes #4. #11 will be done once we have more content on the page and can be tested. on narrower screens, there is a burger that, on click event, opens a drawer with the navigation links inside. components have been split into subcomponents to reduce markup - this includes having separate components for CTA button to book free trial and the nav links themselves. routing has been implemented, though the 404 page has not been created yet (that will be done in #2) so the classes links don't go anywhere. added proptypes, though the isNarrowScreen boolean is undefined on initial load, but quickly gets a value (checked using console.log). this boolean is defined using mantine's useMediaQuery hook where a 'narrow' screen is one that is <= 739px. I got this value from spot testing.