feat: build mvp homepage.
fixes #1 in its core subparts. i have not included the contact form (#7) as we are presuming there will be a contact page (#14) that has the utility of providing parking instructions and a map of the gym's location. the hero, below-the-fold, classes, and 'why choose us' sections have been built. #11 will be done in a future iteration and be applied to all pages.