feat: on wider screens, once user scrolls down, the header becomes slightly transparent and reduces in size. we use a shorter (vertically) logo. fix: on very narrow screens, fix the free trial cta button pushing left into the EQA logo.
on narrower screens, there is sufficient height-wise page real estate to not justify having to alter the header height. on wider screens, the vertical space is precious, so having reduction in size of header helps the user to see more of the main sections of the page. the reduction in opacity (more transparent) header similarly allows the user to be able to see what they've just scrolled past. i've leveraged mantine's hooks to avoid having to write my own. the header is also no longer sticky but fixed to not have this faint, thin, fully transparent line at the top of the page once we scroll past a certain point. fixes #11.