feat: build adults and kids classes pages. ensure responsiveness.
i intended to share the page structure for classes (and potentially even to timetable pages and maybe contact page) with the above-the-fold, sections broken up by colours, and the cta section at the bottom in bright colours to push for conversion. this is why there is a css module file that handles almost all of the shared styles across the pages, and some that are specific to the adults pages. those specific styling composes shared styles and make slight tweaks. i know i need to refactor these pages, which i'll handle in #25. the women's page (#21) is requiring some confirmation on the intention of business owners which will flow on to how the page will be structured, and thus the copy, and potentially even a separately embedded form. This completes two subtasks (fixes #19, fixes #20) of #6.