feat: build contact page, which includes a contact form, and a map of the location. make responsive.
fixes #7. EmailJS-integrated. When someone fills in the contact form and sends it, equinox admin team will get an alert with all the details in the contact form that was filled out. there is no email sent to the user to confirm it has been received yet, but this can always be built in the future. when this whole website is released into prod, we will do #35, but right now, we may need to keep it whilst still in dev environment. I didn't want to pay Google API key, so I made my own screenshot with the location pin of the gym, which once clicked, takes the user to Google Maps anyway, and then the user can find directions from their location etc. handle loading and sent states of the contact form as well such that when the form is being sent, all fields are disabled. after successfully sending, an article is shown to advise the user that the message has been sent