feat: add timetable page with integrated clubworx timetable. ensure responsiveness. before the clubworx timetable iframe jsx element is loaded, a skeleton placeholder is shown.
fixes #32. the design and styling here aligns with that of Legacy BJJ, which also uses Clubworx. also add the above the fold section which aligns in styling with the classes pages, and the cta section at the bottom (orange semi-circle with cta to book a free trial. this is part of #15.