feat: add error page for when a page is not found.
make responsive - becomes a single column on narrower screens, but 2 columns on wider screens. use a cute image (like amazon does with the sad puppy) to try and not make users mad. add copy. ensure header and footer are visible and provide a link to return to home page. fixes #2