refactor: CTA button is a separate component that passes in props (fixes #33). feat: CTA button now links to Clubworx for user to sign waiver. refactor: above the fold section on most sub pages (except contact at the moment) is its own component (fixes #45)
this helps keep the code DRY, especially for certain props for Mantine components that repeated, like the colour and radius, as well as the CSS classes and JSX markup for the above the fold sections across the pages. This improves maintainability greatly, as well as making the pages themselves more readable.