refactor(Homepage): the cards representing classes is now an <article> rather than itself also being a React Router Link. this fixes console error where an anchor should not be nested within an anchor (#27).
opted for the options listed in #27 comments - kept hover effects for responsiveness, but removed the cursor being a clicked one (to avoid frustration when they can't actually click the entire article card). the alternative option of rendering based on whether it is a mobile or tablet is hard to execute without potentially routing (eg adding a 'm.' prefix before the domain), which is overkill for just this single feature. we also remove the need to have a copy to advise user what to do on wider screens (i.e. 'click the card to find out more' or something like that). this is better because UI should be self-evident and intuitive as much as possible, without need to provide further instructions.