fix: console errors (#27) and various responsiveness issues (#46) (#51)
fixes #27 and fixes #46

* fix(Footer): ensure every child JSX element has a React key attribute

this was a console error in dev in issue #27

* style: increase delay of start of animation for underlining the 'everyone' copy in the hero section of home page

* style: on narrower screens, the classes cards on home page have less padding. reduce margin-top for the 'learn more' button so that it doesn't feel so 'spacey' on narrower screens.

this takes better take advantage of more limited page real estate. 'learn more' button's margin-top reduction helps make the card look less spaced out, especially for the 'kids' card which has less copy than the other two. this is related to #46.

* style(Homepage): remove padding-bottom of the 'why choose us' section on home page

there is too much space on smaller screens now that there is the cta section - this section (in bright orange) should be within view when the user is seeing the 'why choose us' section to maximise conversion. this goes to issue #46

* style(Footer): increase size of instagram icon, especially on smaller screens.

this goes to issue #46. the instagram icon is actually a clickable anchor, which means it should (design-wise) be more important than the icons that are merely decorative for the other parts of the footer for the various contact methods.

* fix(ContactPage): ensure every JSX markup being generated from .map has a React-required key attribute.

this fixes a console error - #27

* refactor(Homepage): the cards representing classes is now an <article> rather than itself also being a React Router Link. this fixes console error where an anchor should not be nested within an anchor (#27).

opted for the options listed in #27 comments - kept hover effects for responsiveness, but removed the cursor being a clicked one (to avoid frustration when they can't actually click the entire article card). the alternative option of rendering based on whether it is a mobile or tablet is hard to execute without potentially routing (eg adding a 'm.' prefix before the domain), which is overkill for just this single feature. we also remove the need to have a copy to advise user what to do on wider screens (i.e. 'click the card to find out more' or something like that). this is better because UI should be self-evident and intuitive as much as possible, without need to provide further instructions.