fix(Header): alignment of CTA in header was previously not being applied. now it is justified to the right-most part of header.
this was a bug that was introduced in c9c8d29. the header is the only instance of CTA button component that has a classname attached to it to give it custom CSS styling. however, the class was applied to the child rather than the parent, so it wasn't being applied.