feat: all images have skeleton while loading (fixes #38). style: default border radius for mantine components is 'xl'. fix/chore: correct mobile number in Footer and Contact Page (fixes #55).
for images with skeleton wrapper, to keep code DRY, there is now a component that wraps a Mantine Skeleton component around an <img> tag. the component takes in the image path, CSS class name for custom styling, and alt text for accessibility considerations (#12). using a default border radius removes the need for many of the components from Mantine to have a separate prop for border radius being XL, which also keeps the code DRY.