chore: replace image for kids' muay thai classes, and 'why choose us' section image on home page
it is much better to use non-stock photos - also saves having to give attribution. this aligns with the rest of the images. use a more high-quality image for the home page because the original was a bit blurry (and was also not actually taken at EQA gym). ensure responsiveness of images with positioning. fixes #57 for now.