implement changes from design review (#60)
* style: the 'book a free trial' cta button in the cta section expands on hover.

it is not very accessible to not have a hover effect on the white button (by default, mantine doesn't seem to enable this).

* style: increase size of 'ready to try a class' text and cta button at the bottom of pages. ensure responsiveness.

makes more sense to try and maximise conversions here.

* style(homepage): use a darker orange colour for the 'learn more' button for class cards.

improves readability and thus accessibility.

* style(homepage): update styling of badge at the top of each section. reduce space between badge and heading.

more legible, and does not clash with secondary button (clickable, whereas this badge is not clickable).

* style(above the fold): secondary buttons now have a darker shade of orange with a transparent background

this is more contemporary styling. avoids colour contrast issues due to the darker shade used.

* style(Header): classes menu also has a hover effect.

aligns with the other links, even though it itself is not a page.

* chore: replace image for kids' muay thai classes, and 'why choose us' section image on home page

it is much better to use non-stock photos - also saves having to give attribution. this aligns with the rest of the images. use a more high-quality image for the home page because the original was a bit blurry (and was also not actually taken at EQA gym). ensure responsiveness of images with positioning. fixes #57 for now.

* style(Header): only the background colour of the header reduces in opacity when scrolled down

this avoids the text being less legible, the cta button being impacted, and also the dropdown for the classes (Mantine Menu component) from also inheriting this opacity reduction. the user might want to navigate away from page after scrolling down already.

* chore(Footer): add book a free trial link in footer

* fix(Homepage): remove overflow of background colour at the bottom of banner image / banner carousel in hero section. reduce gap in flexbox for text above the hero image. increase padding on top and bottom of hero.

noticeable and looks a bit odd, so glad we got rid of the extra bit of background. this does mean more markup in jsx and css by adding in the extra parent container. adding more spacing gives it more 'breathing room' visually.