fix(Footer): ensure gap in footer on narrower screens is consistent (fixes #62). chore(KidsClassesPage): add 'active kids' certification in copy (fixes #59).
refactored above the fold section component to take in jsx as an optional prop to be rendered under the heading and above the cta buttons. the inconsistent gap in footer was caused by adding an additional link in the 'browse' column of footer, added in PR #60, to allow users to book a free trial more easily. on narrower screens now, we use a flexbox (column) to stack them, rather than maintaining display: grid with a single column. i tried to add the logo into the footer but it looked ugly, so opted to have it in the above-the-fold section of the kids page. this is because the active kids vouchers are only really used for kids classes anyway. added a link to service nsw's website so that parents/guardians can look up more info. i confirmed via the portal as well that eqa is already a registered provider.