style(HomePage): re-align all elements of classes cards on homepage. remove hover effect on cards. increase delay before beginning the underlining of text in hero, and speed up animation.
the misalignment of text was caused by copy changes in #69. this fixes #72. the hover effect might be confusing because the cards themselves are not clickable. removing hover also enables #31 to not look so odd (where a 3-column section, on top of each other, both are hovered), though #31 would just be for aesthetic and not function (we can have the default cursor to avoid user confusion and frustration).